etl/requirements.txt
pandas 
pymysql
sqlalchemy
python-dotenv
cryptography